requirements.txt
Flask
pandas
scikit-learn
google-generativeai
flask-cors
gunicorn
